requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-lichess-sdk==1.1.0.7
async-timeout==4.0.3
attrs==24.2.0
Authlib==1.3.2
berserk==0.13.2
blessed==1.20.0
blinker==1.8.2
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chess==1.11.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.3
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
Deprecated==1.2.14
discord.py==2.4.0
docker==7.1.0
evdev==1.7.1
exceptiongroup==1.2.2
fastapi==0.115.2
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.5.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
keyboard==0.13.5
kiwisolver==1.4.7
libtmux==0.37.0
lockfile==0.12.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
multidict==6.1.0
ndjson==0.3.1
numpy==2.1.2
oauthlib==3.2.2
packaging==24.1
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pynput==1.7.7
pyparsing==3.2.0
PyTermGUI==7.7.2
python-chess==1.999
python-crontab==3.2.0
python-daemon==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-xlib==0.33
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.2
six==1.16.0
sniffio==1.3.1
starlette==0.40.0
tinycss2==1.4.0
tmuxp==1.47.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
Wand==0.6.13
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
wrapt==1.16.0
yarl==1.17.0
